Check if GC is marking in art::ReadBarrier::Barrier
Check art::Thread::GetIsGcMarking in art::ReadBarrier::Barrier
before checking whether the holder of the reference is gray.
The former check is expected to be cheaper than the latter, and
thus is expected to make art::ReadBarrier::Barrier calls (mainly
from the interpreter) a bit faster when the GC is not marking.

Test: art/test.py
Bug: 12687968
Change-Id: I97deeffb657e33e7d6b975b46df69805021a4c0e
Signed-off-by: mydongistiny <jaysonedson@gmail.com>